move currency and fee constants into shared area for re-use
