+++ parachains/runtimes/coretime/coretime-kusama/src/constants.rs
@@ -1,109 +1,109 @@
-// Copyright (C) 2021 Parity Technologies (UK) Ltd.
-// SPDX-License-Identifier: Apache-2.0
+// // Copyright (C) 2021 Parity Technologies (UK) Ltd.
+// // SPDX-License-Identifier: Apache-2.0
 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// // Licensed under the Apache License, Version 2.0 (the "License");
+// // you may not use this file except in compliance with the License.
+// // You may obtain a copy of the License at
+// //
+// // 	http://www.apache.org/licenses/LICENSE-2.0
+// //
+// // Unless required by applicable law or agreed to in writing, software
+// // distributed under the License is distributed on an "AS IS" BASIS,
+// // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// // See the License for the specific language governing permissions and
+// // limitations under the License.
 
-pub mod currency {
-	use kusama_runtime_constants as constants;
-	use polkadot_core_primitives::Balance;
+// pub mod currency {
+// 	use kusama_runtime_constants as constants;
+// 	use polkadot_core_primitives::Balance;
 
-	/// The existential deposit. Set to 1/10 of its parent Relay Chain.
-	pub const EXISTENTIAL_DEPOSIT: Balance = constants::currency::EXISTENTIAL_DEPOSIT / 10;
+// 	/// The existential deposit. Set to 1/10 of its parent Relay Chain.
+// 	pub const EXISTENTIAL_DEPOSIT: Balance = constants::currency::EXISTENTIAL_DEPOSIT / 10;
 
-	pub const UNITS: Balance = constants::currency::UNITS;
-	pub const CENTS: Balance = constants::currency::CENTS;
-	pub const GRAND: Balance = constants::currency::GRAND;
-	pub const MILLICENTS: Balance = constants::currency::MILLICENTS;
+// 	pub const UNITS: Balance = constants::currency::UNITS;
+// 	pub const CENTS: Balance = constants::currency::CENTS;
+// 	pub const GRAND: Balance = constants::currency::GRAND;
+// 	pub const MILLICENTS: Balance = constants::currency::MILLICENTS;
 
-	pub const fn deposit(items: u32, bytes: u32) -> Balance {
-		// map to 1/100 of what the kusama relay chain charges (v9020)
-		constants::currency::deposit(items, bytes) / 100
-	}
-}
+// 	pub const fn deposit(items: u32, bytes: u32) -> Balance {
+// 		// map to 1/100 of what the kusama relay chain charges (v9020)
+// 		constants::currency::deposit(items, bytes) / 100
+// 	}
+// }
 
-/// Fee-related.
-pub mod fee {
-	use frame_support::{
-		pallet_prelude::Weight,
-		weights::{
-			constants::ExtrinsicBaseWeight, FeePolynomial, WeightToFeeCoefficient,
-			WeightToFeeCoefficients, WeightToFeePolynomial,
-		},
-	};
-	use polkadot_core_primitives::Balance;
-	use smallvec::smallvec;
-	pub use sp_runtime::Perbill;
+// /// Fee-related.
+// pub mod fee {
+// 	use frame_support::{
+// 		pallet_prelude::Weight,
+// 		weights::{
+// 			constants::ExtrinsicBaseWeight, FeePolynomial, WeightToFeeCoefficient,
+// 			WeightToFeeCoefficients, WeightToFeePolynomial,
+// 		},
+// 	};
+// 	use polkadot_core_primitives::Balance;
+// 	use smallvec::smallvec;
+// 	pub use sp_runtime::Perbill;
 
-	/// The block saturation level. Fees will be updates based on this value.
-	pub const TARGET_BLOCK_FULLNESS: Perbill = Perbill::from_percent(25);
+// 	/// The block saturation level. Fees will be updates based on this value.
+// 	pub const TARGET_BLOCK_FULLNESS: Perbill = Perbill::from_percent(25);
 
-	/// Handles converting a weight scalar to a fee value, based on the scale and granularity of the
-	/// node's balance type.
-	///
-	/// This should typically create a mapping between the following ranges:
-	///   - [0, MAXIMUM_BLOCK_WEIGHT]
-	///   - [Balance::min, Balance::max]
-	///
-	/// Yet, it can be used for any other sort of change to weight-fee. Some examples being:
-	///   - Setting it to `0` will essentially disable the weight fee.
-	///   - Setting it to `1` will cause the literal `#[weight = x]` values to be charged.
-	pub struct WeightToFee;
-	impl frame_support::weights::WeightToFee for WeightToFee {
-		type Balance = Balance;
+// 	/// Handles converting a weight scalar to a fee value, based on the scale and granularity of the
+// 	/// node's balance type.
+// 	///
+// 	/// This should typically create a mapping between the following ranges:
+// 	///   - [0, MAXIMUM_BLOCK_WEIGHT]
+// 	///   - [Balance::min, Balance::max]
+// 	///
+// 	/// Yet, it can be used for any other sort of change to weight-fee. Some examples being:
+// 	///   - Setting it to `0` will essentially disable the weight fee.
+// 	///   - Setting it to `1` will cause the literal `#[weight = x]` values to be charged.
+// 	pub struct WeightToFee;
+// 	impl frame_support::weights::WeightToFee for WeightToFee {
+// 		type Balance = Balance;
 
-		fn weight_to_fee(weight: &Weight) -> Self::Balance {
-			let time_poly: FeePolynomial<Balance> = RefTimeToFee::polynomial().into();
-			let proof_poly: FeePolynomial<Balance> = ProofSizeToFee::polynomial().into();
+// 		fn weight_to_fee(weight: &Weight) -> Self::Balance {
+// 			let time_poly: FeePolynomial<Balance> = RefTimeToFee::polynomial().into();
+// 			let proof_poly: FeePolynomial<Balance> = ProofSizeToFee::polynomial().into();
 
-			// Take the maximum instead of the sum to charge by the more scarce resource.
-			time_poly.eval(weight.ref_time()).max(proof_poly.eval(weight.proof_size()))
-		}
-	}
+// 			// Take the maximum instead of the sum to charge by the more scarce resource.
+// 			time_poly.eval(weight.ref_time()).max(proof_poly.eval(weight.proof_size()))
+// 		}
+// 	}
 
-	/// Maps the reference time component of `Weight` to a fee.
-	pub struct RefTimeToFee;
-	impl WeightToFeePolynomial for RefTimeToFee {
-		type Balance = Balance;
-		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
-			// in Kusama, extrinsic base weight (smallest non-zero weight) is mapped to 1/10 CENT:
-			// in Asset Hub, we map to 1/10 of that, or 1/100 CENT
-			let p = super::currency::CENTS;
-			let q = 100 * Balance::from(ExtrinsicBaseWeight::get().ref_time());
+// 	/// Maps the reference time component of `Weight` to a fee.
+// 	pub struct RefTimeToFee;
+// 	impl WeightToFeePolynomial for RefTimeToFee {
+// 		type Balance = Balance;
+// 		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
+// 			// in Kusama, extrinsic base weight (smallest non-zero weight) is mapped to 1/10 CENT:
+// 			// in Asset Hub, we map to 1/10 of that, or 1/100 CENT
+// 			let p = super::currency::CENTS;
+// 			let q = 100 * Balance::from(ExtrinsicBaseWeight::get().ref_time());
 
-			smallvec![WeightToFeeCoefficient {
-				degree: 1,
-				negative: false,
-				coeff_frac: Perbill::from_rational(p % q, q),
-				coeff_integer: p / q,
-			}]
-		}
-	}
+// 			smallvec![WeightToFeeCoefficient {
+// 				degree: 1,
+// 				negative: false,
+// 				coeff_frac: Perbill::from_rational(p % q, q),
+// 				coeff_integer: p / q,
+// 			}]
+// 		}
+// 	}
 
-	/// Maps the proof size component of `Weight` to a fee.
-	pub struct ProofSizeToFee;
-	impl WeightToFeePolynomial for ProofSizeToFee {
-		type Balance = Balance;
-		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
-			// Map 10kb proof to 1 CENT.
-			let p = super::currency::CENTS;
-			let q = 10_000;
+// 	/// Maps the proof size component of `Weight` to a fee.
+// 	pub struct ProofSizeToFee;
+// 	impl WeightToFeePolynomial for ProofSizeToFee {
+// 		type Balance = Balance;
+// 		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
+// 			// Map 10kb proof to 1 CENT.
+// 			let p = super::currency::CENTS;
+// 			let q = 10_000;
 
-			smallvec![WeightToFeeCoefficient {
-				degree: 1,
-				negative: false,
-				coeff_frac: Perbill::from_rational(p % q, q),
-				coeff_integer: p / q,
-			}]
-		}
-	}
-}
+// 			smallvec![WeightToFeeCoefficient {
+// 				degree: 1,
+// 				negative: false,
+// 				coeff_frac: Perbill::from_rational(p % q, q),
+// 				coeff_integer: p / q,
+// 			}]
+// 		}
+// 	}
+// }
