+++ parachains/runtimes/coretime/coretime-polkadot/src/constants.rs
@@ -13,97 +13,97 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod currency {
-	use polkadot_core_primitives::Balance;
-	use polkadot_runtime_constants as constants;
+// pub mod currency {
+// 	use polkadot_core_primitives::Balance;
+// 	use polkadot_runtime_constants as constants;
 
-	/// The existential deposit. Set to 1/10 of its parent Relay Chain.
-	pub const EXISTENTIAL_DEPOSIT: Balance = constants::currency::EXISTENTIAL_DEPOSIT / 10;
+// 	/// The existential deposit. Set to 1/10 of its parent Relay Chain.
+// 	pub const EXISTENTIAL_DEPOSIT: Balance = constants::currency::EXISTENTIAL_DEPOSIT / 10;
 
-	pub const UNITS: Balance = constants::currency::UNITS;
-	pub const DOLLARS: Balance = constants::currency::DOLLARS;
-	pub const CENTS: Balance = constants::currency::CENTS;
-	pub const MILLICENTS: Balance = constants::currency::MILLICENTS;
+// 	pub const UNITS: Balance = constants::currency::UNITS;
+// 	pub const DOLLARS: Balance = constants::currency::DOLLARS;
+// 	pub const CENTS: Balance = constants::currency::CENTS;
+// 	pub const MILLICENTS: Balance = constants::currency::MILLICENTS;
 
-	pub const fn deposit(items: u32, bytes: u32) -> Balance {
-		// 1/100 of Polkadot
-		constants::currency::deposit(items, bytes) / 100
-	}
-}
+// 	pub const fn deposit(items: u32, bytes: u32) -> Balance {
+// 		// 1/100 of Polkadot
+// 		constants::currency::deposit(items, bytes) / 100
+// 	}
+// }
 
-/// Fee-related.
-pub mod fee {
-	use frame_support::{
-		pallet_prelude::Weight,
-		weights::{
-			constants::ExtrinsicBaseWeight, FeePolynomial, WeightToFeeCoefficient,
-			WeightToFeeCoefficients, WeightToFeePolynomial,
-		},
-	};
-	use polkadot_core_primitives::Balance;
-	use smallvec::smallvec;
-	pub use sp_runtime::Perbill;
+// /// Fee-related.
+// pub mod fee {
+// 	use frame_support::{
+// 		pallet_prelude::Weight,
+// 		weights::{
+// 			constants::ExtrinsicBaseWeight, FeePolynomial, WeightToFeeCoefficient,
+// 			WeightToFeeCoefficients, WeightToFeePolynomial,
+// 		},
+// 	};
+// 	use polkadot_core_primitives::Balance;
+// 	use smallvec::smallvec;
+// 	pub use sp_runtime::Perbill;
 
-	/// The block saturation level. Fees will be updates based on this value.
-	pub const TARGET_BLOCK_FULLNESS: Perbill = Perbill::from_percent(25);
+// 	/// The block saturation level. Fees will be updates based on this value.
+// 	pub const TARGET_BLOCK_FULLNESS: Perbill = Perbill::from_percent(25);
 
-	/// Handles converting a weight scalar to a fee value, based on the scale and granularity of the
-	/// node's balance type.
-	///
-	/// This should typically create a mapping between the following ranges:
-	///   - [0, MAXIMUM_BLOCK_WEIGHT]
-	///   - [Balance::min, Balance::max]
-	///
-	/// Yet, it can be used for any other sort of change to weight-fee. Some examples being:
-	///   - Setting it to `0` will essentially disable the weight fee.
-	///   - Setting it to `1` will cause the literal `#[weight = x]` values to be charged.
-	pub struct WeightToFee;
-	impl frame_support::weights::WeightToFee for WeightToFee {
-		type Balance = Balance;
+// 	/// Handles converting a weight scalar to a fee value, based on the scale and granularity of the
+// 	/// node's balance type.
+// 	///
+// 	/// This should typically create a mapping between the following ranges:
+// 	///   - [0, MAXIMUM_BLOCK_WEIGHT]
+// 	///   - [Balance::min, Balance::max]
+// 	///
+// 	/// Yet, it can be used for any other sort of change to weight-fee. Some examples being:
+// 	///   - Setting it to `0` will essentially disable the weight fee.
+// 	///   - Setting it to `1` will cause the literal `#[weight = x]` values to be charged.
+// 	pub struct WeightToFee;
+// 	impl frame_support::weights::WeightToFee for WeightToFee {
+// 		type Balance = Balance;
 
-		fn weight_to_fee(weight: &Weight) -> Self::Balance {
-			let time_poly: FeePolynomial<Balance> = RefTimeToFee::polynomial().into();
-			let proof_poly: FeePolynomial<Balance> = ProofSizeToFee::polynomial().into();
+// 		fn weight_to_fee(weight: &Weight) -> Self::Balance {
+// 			let time_poly: FeePolynomial<Balance> = RefTimeToFee::polynomial().into();
+// 			let proof_poly: FeePolynomial<Balance> = ProofSizeToFee::polynomial().into();
 
-			// Take the maximum instead of the sum to charge by the more scarce resource.
-			time_poly.eval(weight.ref_time()).max(proof_poly.eval(weight.proof_size()))
-		}
-	}
+// 			// Take the maximum instead of the sum to charge by the more scarce resource.
+// 			time_poly.eval(weight.ref_time()).max(proof_poly.eval(weight.proof_size()))
+// 		}
+// 	}
 
-	/// Maps the reference time component of `Weight` to a fee.
-	pub struct RefTimeToFee;
-	impl WeightToFeePolynomial for RefTimeToFee {
-		type Balance = Balance;
-		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
-			// in Polkadot, extrinsic base weight (smallest non-zero weight) is mapped to 1/10 CENT:
-			// in Asset Hub, we map to 1/10 of that, or 1/100 CENT
-			let p = super::currency::CENTS;
-			let q = 100 * Balance::from(ExtrinsicBaseWeight::get().ref_time());
+// 	/// Maps the reference time component of `Weight` to a fee.
+// 	pub struct RefTimeToFee;
+// 	impl WeightToFeePolynomial for RefTimeToFee {
+// 		type Balance = Balance;
+// 		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
+// 			// in Polkadot, extrinsic base weight (smallest non-zero weight) is mapped to 1/10 CENT:
+// 			// in Asset Hub, we map to 1/10 of that, or 1/100 CENT
+// 			let p = super::currency::CENTS;
+// 			let q = 100 * Balance::from(ExtrinsicBaseWeight::get().ref_time());
 
-			smallvec![WeightToFeeCoefficient {
-				degree: 1,
-				negative: false,
-				coeff_frac: Perbill::from_rational(p % q, q),
-				coeff_integer: p / q,
-			}]
-		}
-	}
+// 			smallvec![WeightToFeeCoefficient {
+// 				degree: 1,
+// 				negative: false,
+// 				coeff_frac: Perbill::from_rational(p % q, q),
+// 				coeff_integer: p / q,
+// 			}]
+// 		}
+// 	}
 
-	/// Maps the proof size component of `Weight` to a fee.
-	pub struct ProofSizeToFee;
-	impl WeightToFeePolynomial for ProofSizeToFee {
-		type Balance = Balance;
-		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
-			// Map 10kb proof to 1 CENT.
-			let p = super::currency::CENTS;
-			let q = 10_000;
+// 	/// Maps the proof size component of `Weight` to a fee.
+// 	pub struct ProofSizeToFee;
+// 	impl WeightToFeePolynomial for ProofSizeToFee {
+// 		type Balance = Balance;
+// 		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
+// 			// Map 10kb proof to 1 CENT.
+// 			let p = super::currency::CENTS;
+// 			let q = 10_000;
 
-			smallvec![WeightToFeeCoefficient {
-				degree: 1,
-				negative: false,
-				coeff_frac: Perbill::from_rational(p % q, q),
-				coeff_integer: p / q,
-			}]
-		}
-	}
-}
+// 			smallvec![WeightToFeeCoefficient {
+// 				degree: 1,
+// 				negative: false,
+// 				coeff_frac: Perbill::from_rational(p % q, q),
+// 				coeff_integer: p / q,
+// 			}]
+// 		}
+// 	}
+// }
